feat: add setup-gradle-pkg action.yml file
This commit adds a new file `action.yml` under `.github/actions/setup-gradle-pkg` directory. The `action.yml` file is used to define a GitHub Action called "Setup Maven Github Pkg". This action sets up Maven package for the repository.

The action runs using the "composite" strategy and consists of a single step. The step runs a bash command to create a `~/.gradle/init.gradle.kts` file with the necessary configuration for Maven package setup. The configuration includes adding repositories and credentials for the Maven package hosted on `https://maven.pkg.github.com/komune-io/fixers`. The credentials are obtained from the `PKG_MAVEN_USERNAME` and `PKG_MAVEN_TOKEN` environment variables.

This new action is added to enable the setup of Maven package for the repository, allowing for easier dependency management and package publishing.

Note: There is no newline at the end of the file.